Improve annotation UI visibility, input validation, and docs
- Enhanced point markers with outer glow for visibility on any background
- Added white outer stroke to polygon contour and box drawings
- Increased polygon close radius from 14px to 20px for easier closing
- Added prompt list validation (type check, max 50 limit) to /annotate route
- Replaced error re-raise with generic error message to prevent info leaks
- Fixed packbits consistency (removed unnecessary .astype(np.uint8))
- Updated README and CLAUDE.md with manual annotation feature documentation

+++ segmenter.py
@@ -551,7 +551,7 @@ class ImageSegmenter:
             # Save compact mask crop
             np.savez_compressed(
                 os.path.join(masks_dir, f"{idx}.npz"),
-                mask=np.packbits(mask_crop.astype(np.uint8)),
+                mask=np.packbits(mask_crop),
                 shape=np.array(mask_crop.shape, dtype=np.int32),
             )
